refactor: 추상클래스 메소드 이동
