+++ src/test/java/view/ResultViewTest.java
@@ -1,5 +1,7 @@
 package view;
 
+import static org.junit.Assert.assertEquals;
+
 import org.junit.Test;
 
 import model.Game;
@@ -8,9 +10,8 @@ public class ResultViewTest {
 	
 	@Test
 	public void resultViewTest() {
+		
 		Game game = new Game(10);
-		game.addPlayer("ABC");
-		game.addPlayer("ABC");
 		game.addPlayer("ABC");
 		game.gameStart();
 		game.shot(10);
@@ -25,37 +26,12 @@ public class ResultViewTest {
 		game.shot(2);
 		game.shot(2);
 		game.shot(2);
-		game.shot(2);
-		
+		game.shot(8);
 		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(10);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
-		game.shot(2);
 		
 		ResultView resultView = new ResultView();
 		resultView.currentResultView(game);
+		
+		assertEquals(214, game.getTotalScore(0));
 	}
 }
